+++ src/components/AlertDialog.tsx
@@ -3,7 +3,7 @@ import Typography, {TypographyVariant} from "./Typography"
 import Button, {ButtonVariant} from "./Button"
 import {IoCloseCircle} from "react-icons/io5"
 import useTranslation from "@/hooks/util/useTranslation"
-import {LangKey} from "@/utils/lang"
+import {LangKey} from "@/lang/allKeys"
 
 export type AlertDialogProps = {
   title: string
