+++ src/components/EventGroupMessage.tsx
@@ -12,7 +12,7 @@ import {twMerge} from "tailwind-merge"
 import Typography from "./Typography"
 import {IoCube} from "react-icons/io5"
 import useTranslation from "@/hooks/util/useTranslation"
-import {LangKey} from "@/utils/lang"
+import {LangKey} from "@/lang/allKeys"
 
 export enum EventShortenerType {
   EqualInfo = "EqualInfo",
