+++ src/components/UserCard.tsx
@@ -6,7 +6,7 @@ import UserProfile, {
   type UserProfileProps as UserProfileProperties,
 } from "./UserProfile"
 import Typography, {TypographyVariant} from "./Typography"
-import {LangKey} from "@/utils/lang"
+import {LangKey} from "@/lang/allKeys"
 import useTranslation from "@/hooks/util/useTranslation"
 
 export type UserCardProps = {
