+++ src/components/VideoPreview.tsx
@@ -3,7 +3,7 @@ import Button, {ButtonVariant} from "./Button"
 import Typography, {TypographyVariant} from "./Typography"
 import {type FC} from "react"
 import useTranslation from "@/hooks/util/useTranslation"
-import {LangKey} from "@/utils/lang"
+import {LangKey} from "@/lang/allKeys"
 
 export type VideoPreviewProps = {
   videoSrc: string
