+++ src/containers/NavigationSection/NotificationsTray.tsx
@@ -7,7 +7,7 @@ import {
   HoverCardTrigger,
 } from "@/components/ui/hover-card"
 import useTranslation from "@/hooks/util/useTranslation"
-import {LangKey} from "@/utils/lang"
+import {LangKey} from "@/lang/allKeys"
 import {useMemo, type FC} from "react"
 
 const sortNotificationsByReadState = (
