+++ src/containers/RoomContainer/ChatInput.tsx
@@ -10,7 +10,7 @@ import {Button} from "@/components/ui/button"
 import useTooltip from "@/hooks/util/useTooltip"
 import useDebounced from "@/hooks/util/useDebounced"
 import useTranslation from "@/hooks/util/useTranslation"
-import {LangKey} from "@/utils/lang"
+import {LangKey} from "@/lang/allKeys"
 
 export type MessageSendRequest = {
   roomId: string
