+++ src/hooks/util/useTranslation.ts
@@ -1,5 +1,11 @@
-import {type LangKey, type Lang, t, setDefaultLang} from "@/utils/lang"
-import {create} from "zustand/react"
+import {type LangKey} from "@/lang/allKeys"
+import {
+  getDefaultLang,
+  type Lang,
+  setDefaultLang,
+  translate,
+} from "@/utils/lang"
+import {create} from "zustand"
 
 type TranslationStore = {
   lang: Lang
@@ -7,15 +13,15 @@ type TranslationStore = {
   t: (key: LangKey, ...args: string[]) => string
 }
 
-const useTranslation = create<TranslationStore>(set => ({
-  lang: "en",
+const useTranslation = create<TranslationStore>((set, get) => ({
+  lang: getDefaultLang(),
   setActiveLang(lang) {
     setDefaultLang(lang)
 
     set(_state => ({lang}))
   },
   t(key, ...args) {
-    return t(key, ...args)
+    return translate(key, get().lang, ...args)
   },
 }))
